requirements.txt
opencv-python
numpy
matplotlib
ultralytics
scikit-learn
yellowbrick
PyYAML
ipykernel